refactor: frame prop types and dom utils
